movie-django-back/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
Django==3.2.9
django-cors-headers==3.10.0
django-dotenv==1.4.2
django-extensions==3.1.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
Pillow==8.4.0
PyJWT==1.7.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.7
